Re-add all the legacy behavior.  This fixes #15, gets us to a2.  Need to fix the docs.
It was surprisingly easy to do.  Main issue was to be clever about waiting for the first UMAPI call
in order to find the organization ID before creating the connection.  Moved all the legacy stuff
to a separate module called legacy so that users have to explicitly ask for it as umapi_client.legacy.
Also had to make connection.make_call return the result, so that it could be available to the UMAPI
client.  This turned out to be a better idea all around, since now errors all have the result in them
for inspection by the caller.